refactor: ベンチメンバーコンポーネントを独立化し、追加の買い物リストを削除
